requirements.txt
Flask
numpy
torch
torchvision
Pillow
scikit-learn
tqdm